refactor: rename root-component to root-level
