+++ packages/spectacular/src/lib/root-level/spectacular-app.component.ts
