+++ packages/spectacular/src/lib/root-level/spectacular-app.module.ts
