+++ packages/spectacular/src/lib/root-level/spectacular-app.scam.ts
